класс статистики сделан
