+++ src/js/analysis/analysis.js
@@ -2,24 +2,21 @@ import "../../pages/analysis.css";
 import "../../../node_modules/normalize.css/normalize.css"
 
 import {ASKED, WEEK_NEWS, TITLES, WHISKERS, MONTH, DAYS, NOW, NEWS, QUERY} from '../constants/constants.js';
-import {calcPercents, renderQuery, renderQuantity, renderPercents, renderDays} from '../utils/utils.js';
+import {renderQuery, renderQuantity, renderPercents, renderDays} from '../utils/utils.js';
+import {Statistics} from '../components/Statistics.js';
 
 (function analysisApp() {
 
+    let days = [1,2,3,4,5,6,7];
+    days = days.map((item, index) => {
+        return item = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-index));
+    });
 
-    const days = [];
-    days[0] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-1))
-    days[1] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-2))
-    days[2] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-3))
-    days[3] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-4))
-    days[4] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-5))
-    days[5] = new Date(NOW.getFullYear(), NOW.getMonth(), NOW.getDate()+(-6))
-    days[6] = NOW;
+    const stats = new Statistics(days, NEWS)
 
-
-    renderPercents(calcPercents(days, JSON.parse(localStorage.news)), WHISKERS);
-    renderQuery(ASKED);
-    renderQuantity(NEWS, QUERY, WEEK_NEWS, TITLES);
+    renderPercents(stats.calcPercents(), WHISKERS);
+    renderQuery(ASKED, QUERY);
+    renderQuantity(NEWS, QUERY.toLowerCase(), WEEK_NEWS, TITLES);
     renderDays(MONTH, DAYS, NOW);
-    
+
 }());
